requirements.txt
discord
python-dotenv
google-api-python-client
youtube-dl
pynacl
prettytable
image
